requirements.txt
attrs==19.1.0
-e git+https://github.com/SohierDane/BigQuery_Helper@8615a7f6c1663e7f2d48aa2b32c2dbcb600a440f#egg=bq_helper
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dash==0.40.0
dash-core-components==0.45.0
dash-daq==0.1.4
dash-html-components==0.15.0
dash-renderer==0.21.0
dash-table==3.6.0
decorator==4.4.0
Flask==1.0.2
Flask-Compress==1.4.0
google-api-core==1.8.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.10.0
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.0.1
jupyter-core==4.4.0
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.16.2
pandas==0.24.2
plotly==3.7.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pyrsistent==0.14.11
python-dateutil==2.8.0
python-gflags==3.1.2
pytz==2018.9
requests==2.21.0
retrying==1.3.3
rsa==4.0
six==1.12.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.15.1
